requirements.txt
pandas
numpy
yfinance
sqlalchemy
psycopg2-binary
python-dotenv